+++ R/package.R
@@ -1,0 +1,11 @@
+#' adtseq
+#' 
+#' Identification of Antibody-Derived-Tags in single cell CITE/REAP-seq data
+#' 
+#' @docType package
+#' @author Julian Spagnuolo
+#' @import Rcpp BH RSeqAn Biostrings
+#' @importFrom Rcpp evalCpp
+#' @useDynLib adtseq
+#' @name adtseq
+NULL