+++ src/icgrEncode.cpp
@@ -12,7 +12,15 @@ using namespace Rcpp;
 // Original python implementation:
 // [Yin, C. (2018). Encoding DNA sequences by integer Chaos Game Representation, Journal of Computational Biology](https://arxiv.org/abs/1712.04546)
 // https://github.com/cyinbox/iCGR
-
+//' @title Integer Chaos Game Encoder
+//' @name icgrEncode
+//' @author Julian Spagnuolo
+//' @description This function will encode a DNA sequence as an integer chaos game representation see: 
+//' [Yin, C. (2018). Encoding DNA sequences by integer Chaos Game Representation, Journal of Computational Biology](https://arxiv.org/abs/1712.04546)
+//' for more detail.
+//'
+//' @param seq, a character vector of length = 1
+//' @export
 // [[Rcpp::export]]
 Rcpp::IntegerMatrix icgrEncode(std::string seq) {