+++ NEWS.md
@@ -1,3 +1,8 @@
+# adtseq 0.99.2
+
+* Altered Makevar to enable ZLIB for RSeqan
+* Added documentation for the functions.
+
 # adtseq 0.99.1
 
 * Added Hamming Distance Function hdist.